Big update, many changes, absolute paths
Most of the changes here are for changing a lot of links/paths to stuff from relative to absolute. So that if files move around, it shouldn't be a problem.

_start : defining some constants for easier access. First one, ENV, is detecting localhost (because path differences). Second one is for not having to write $_SERVER['DOCUMENT_ROOT'] everytime something is included/required in PHP (laziness). Third one is reading the ini-file into INI. Might be slightly slower than only reading when needed, but not as important as laziness.

DBconn-class : No more iniFileName-parameter, since it's read in _start (with absolute path, not relative).

html-head : Scripts/css now use absolute path, not relative. ENV is also used here, not just for detecting localhost.

index.php : Absolute path to components.

upload.php : fixes to creating folder (new location for it outside web-root, reading from INI-const, returning array always, not array|bool.) Some absolute paths. Fixed javascript.

Should point out that all the abs paths weren't really needed in this project, but I just wanted to do them because they sounded useful. In the end, it was more work than it was worth, some important paths are still relative (for reasons), and I was being too clever for my own good.